# internal/server/gen/swagger.yaml
components:
  schemas:
...
    github_com_kevin_vigil_internal_httputil.ErrorResponse:
      properties:
        code:
          description: Code 机器可读错误码（可选，非 4xx 通用错误可留空）。
          example: invalid_argument
          type: string
        details:
          description: Details 结构化补充信息（可选，如字段级校验错误）。
        error:
          example: invalid id
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: 业务 API 推荐（唯一声明）方案：HTTP Bearer JWT（/auth/login 换取）。
      scheme: bearer
      type: http
...
info:
  description: |-
    Vigil 告警处置平台 REST API。
    认证：业务 API 要求 Bearer JWT（/auth/login 换取）。鉴权中间件另接受 X-Vigil-User-ID 头作为本地/回退身份（可伪造，仅限受信网络，生产禁用），该回退方案不在 securitySchemes 中声明。
  license:
    name: MIT
  title: Vigil API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /im/{platform}/callback:
    post:
      description: 各平台（feishu/dingtalk/wecom）回调入口：签名校验 → 解析为标准事件 → 派发卡片/命令/@ 处理。公开入口，平台签名鉴权，不走
        RBAC。
      parameters:
      - description: IM 平台（feishu/dingtalk/wecom）
        in: path
        name: platform
        required: true
        schema:
          enum:
          - feishu
          - dingtalk
          - wecom
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_kevin_vigil_internal_httputil.ErrorResponse'
          description: Not Found
      summary: IM 平台回调
      tags:
      - im
...
servers:
- url: /api/v1